Added lexing of comments
Comments now get thier own `TOKEN_COMMENT` type with the content of the
comment placed in the `value` field as a string. This will be useful in
the future when we write tests involving aspen source code, we can write
extra infomation needed to execute the tests directly in the source code
as a comment.