fixed bug where an undefined function could be called
consider:
```
fn a() void {
    b();
}

a();

fn b() void {

}
```

This program would crash the interpreter because `a` references an undefined identifier `b` when it is called.

To fix this we introduce the concept of an "unresolved function" in the type checker. A function is unresolved if either

1. it is currently undefined
2. it references an unresolved function in its body

Here currently means that execution has not reached the function declaration yet

This commit fixes this issue by enforcing that all top level code must not reference unresolved functions.